refactor(shell): extract sidebar drag-drop, panel-resize, setup
Three small isolated extractions from sidebar/index.js:
- drag-drop.js: tab→workspace drop handlers, receives moveTabToWorkspace via DI
- panel-resize.js: resize handle state + logic, exports get/setPanelWidth
- panels/setup.js: setup panel render, receives hideWebviews/safeSetPanelHTML/render via factory

No behavior changes. Cross-module wiring uses explicit dependency injection
to keep the module boundaries clean.
